refactor: add filter to utils
